Getting results with the Linear model
The linear model seems to be giving acceptable
Validity results. The accuracy and cost of recourse
are not the same, though these could be marked up to
the model differences (technically the linear is a LR, so getting
model accuracy the same as the paper's should be possible. Why it isn't
there yet definitly has something to do with the preprocessing of the data)

Roar is having trouble getting recourse for mlp (non-linear) This needs to
be investigated further, given that the application of ROAR is almost the
same as the paper. The only difference between Linear and non-linear model
is the use of LIME. That is something to look at next.