Changed a parameter to the loadData and loadModel
I added a parameter "modified" to load a modified version of the
SBA/german dataset and updated the relevant functions and classes to handle this new parameter.

The results for ROAR seem inline with the paper
in terms of validity. I would try a larger sample of
factuals to be sure.

Sometimes the tests fail, i need to make sure that the
method runs the same every time, interms of factuals and
model training. Right now there isn't consistency in that.