More fine-tune control to PyTorch model training
Added parameters for epochs, batch size, and learning rate to the model loading and training functions to allow better customization of PyTorch neural network training.
This is to help be more inline with the training parameters used in reproducing results for the Probe method.
The results are inline with the paper.

The main difference are in the preprocessing of the datasets,
which in the paper, all the features are treated as continuous features,
while in the current codebase, categorical features are one-hot encoded.
This causes some differences in the final accuracy and results,
but the trends remain the same.